requirements.txt
absl-py==1.4.0
accelerate==0.20.3
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
aspose-cells==24.12.0
aspose-words==24.12.0
Aspose.Slides==24.12.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.0.0
blinker==1.6.3
blis==0.7.11
braceexpand==0.1.7
cached-property==1.5.2
cachetools==5.3.1
catalogue==2.0.10
certifi==2024.2.2
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==7.1.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comet-ml==3.33.11
comm==0.1.4
confection==0.1.5
configobj==5.0.8
configparser==5.3.0
cycler==0.11.0
cymem==2.0.6
data==0.4
debugpy==1.7.0
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.16.0
distlib==0.3.8
docker-pycreds==0.4.0
docx==0.2.4
dulwich==0.21.6
easydict==1.9
einops==0.4.1
en-vectors-web-lg==2.3.0
entrypoints==0.4
et-xmlfile==1.1.0
everett==3.1.0
exceptiongroup==1.2.1
fairscale==0.4.4
faiss==1.5.3
faiss-gpu==1.7.2
fastjsonschema==2.19.1
filelock==3.12.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
funcsigs==1.0.2
future==0.18.3
gensim==4.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.0
google-auth-oauthlib==0.4.6
gql==0.2.0
graphql-core==1.1
greenlet==1.0.0
grpcio==1.58.0
h5py==3.6.0
huggingface-hub==0.16.4
identify==2.5.24
idna==3.3
imageio==2.31.2
imageio-ffmpeg==0.4.9
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipdb==0.13.9
ipykernel==6.16.2
ipython==7.33.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.1.0
JPype1==1.5.0
json5==0.9.16
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_packaging==0.12.3
jupyter_server_fileid==0.9.2
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.10
kaggle==1.6.12
kiwisolver==1.4.5
langcodes==3.3.0
latex==0.7.0
lmdb==1.5.1
lxml==5.3.0
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.1
mkl-service==2.4.0
mmengine==0.10.7
modelscope==1.23.1
moviepy==1.0.3
multidict==6.0.5
murmurhash==1.0.7
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.6.3
ninja==1.11.1.1
nltk==3.7
nodeenv==1.8.0
notebook==6.5.6
notebook_shim==0.2.4
numpy==1.21.6
nvidia-ml-py==12.570.172
nvidia-ml-py3==7.352.0
nvitop==1.4.2
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
open3d==0.10.0.0
openai==0.28.0
opencv-python==4.8.0.74
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openpyxl==3.1.2
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
pathy==0.10.3
pdfminer==20191125
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==4.0.0
playwright==1.9.0
plotly==5.18.0
portalocker==2.7.0
pre-commit==2.21.0
preshed==3.0.6
proglog==0.1.10
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.21.0
pydantic==1.8.2
pydeck==0.8.1b1
pyDeprecate==0.3.1
pyee==10.0.2
Pygments==2.16.1
Pympler==1.0.1
pymssql==2.2.8
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.16
PyQt5-sip==12.13.0
pyrsistent==0.19.3
python-box==6.1.0
python-dateutil==2.8.2
python-docx==1.1.0
python-json-logger==2.0.7
python-magic==0.4.27
python-pptx==0.6.23
python-slugify==8.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==24.0.1
regex==2022.4.24
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
sacremoses==0.0.53
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==1.31.0
setproctitle==1.3.3
shortuuid==1.0.13
shutilwhich==1.1.0
simplejson==3.19.2
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.23.1
subprocess32==3.5.4
tempdir==0.7.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
textrank4zh==0.3
thinc==8.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.4
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
ttach==0.0.3
typer==0.3.2
typing==3.7.4.3
typing_extensions==4.1.1
tzdata==2024.1
tzlocal==4.3.1
ujson==5.3.0
ultralytics==8.0.145
uri-template==1.3.0
urllib3==1.26.17
validators==0.20.0
virtualenv==20.26.1
wandb==0.8.35
wasabi==0.9.1
watchdog==3.0.0
wcwidth==0.2.5
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.15.0
wurlitzer==3.0.3
XlsxWriter==3.2.0
xlwt==1.3.0
y-py==0.6.2
yapf==0.43.0
yarl==1.9.4
ypy-websocket==0.8.4
zipp==3.8.0
